remove ServiceImpl & edit .gitignore
