+++ .env
@@ -1,5 +1,5 @@
 DB_USERNAME=root
-DB_PASSWORD=rlaqudcks715
+DB_PASSWORD=1234
 JWT_SECRET=thisIsMeaningless123StringForJwtSecretKey8888awghiDSjqbwf
 GOOGLE_CLIENT_ID=57872767863-nd9vfqnd173544vv9c028ha8shafj3if.apps.googleusercontent.com
 GOOGLE_CLIENT_SECRET=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
@@ -11,6 +11,6 @@ NAVER_REDIRECT_URI=http://localhost:8080/login/oauth2/code/naver
 NAVER_AUTHORIZATION_URI=https://nid.naver.com/oauth2.0/authorize
 NAVER_TOKEN_URI=https://nid.naver.com/oauth2.0/token
 NAVER_USER_INFO_URI=https://openapi.naver.com/v1/nid/me
-PROJECT_PATH=/Users/byoungchan/Desktop/backend
+PROJECT_PATH=C:/spring_study/backend
 SPRING_REDIS_HOST=redis
 SPRING_REDIS_PORT=6379
