+++ README.md
@@ -1,2 +1,10 @@
 # backend
-웹후크 테스트
+
+도커 실행
+docker-compose up --build
+
+도커 실행 중지
+docker-compose down
+
+수정사항이 반영 안되었을때
+docker-compose up --build --no-cache
